requirements.txt
grpcio==1.41.0
grpcio-tools==1.41.0
kiwisolver==1.3.2
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.4
Pillow==8.4.0
pyparsing==3.0.5
python-dateutil==2.8.2
six==1.16.0
